move view related files into new view folder and lib
