+++ Makefile.in
@@ -83,6 +83,7 @@ DIRECTORY_ARCHIVES=$(DVDPLAYER_ARCHIVES) \
                    xbmc/video/dialogs/videodialogs.a \
                    xbmc/video/video.a \
                    xbmc/video/windows/videowindows.a \
+                   xbmc/view/view.a \
                    xbmc/windowing/windowing.a \
                    xbmc/windows/windows.a \
                    xbmc/xbmc.a \
