+++ xbmc/Makefile.in
@@ -13,8 +13,6 @@ SRCS=Application.cpp \
      GUIInfoManager.cpp \
      GUILargeTextureManager.cpp \
      GUIPassword.cpp \
-     GUIViewControl.cpp \
-     GUIViewState.cpp \
      MediaSource.cpp \
      NfoFile.cpp \
      PasswordManager.cpp \
@@ -30,7 +28,6 @@ SRCS=Application.cpp \
      ThumbnailCache.cpp \
      URL.cpp \
      Util.cpp \
-     ViewDatabase.cpp \
      XBApplicationEx.cpp \
      XBDateTime.cpp \
      xbmc.cpp \
