+++ xbmc/addons/GUIViewStateAddonBrowser.cpp
@@ -21,7 +21,7 @@
 #include "GUIViewStateAddonBrowser.h"
 #include "FileItem.h"
 #include "guilib/Key.h"
-#include "ViewState.h"
+#include "view/ViewState.h"
 #include "settings/Settings.h"
 #include "addons/Addon.h"
 #include "addons/AddonManager.h"
