+++ xbmc/dialogs/GUIDialogFileBrowser.h
@@ -23,7 +23,7 @@
 #include "guilib/GUIDialog.h"
 #include "filesystem/VirtualDirectory.h"
 #include "filesystem/DirectoryHistory.h"
-#include "GUIViewControl.h"
+#include "view/GUIViewControl.h"
 #include "pictures/PictureThumbLoader.h"
 
 class CFileItem;
