+++ xbmc/programs/GUIViewStatePrograms.cpp
@@ -20,7 +20,7 @@
 
 #include "GUIViewStatePrograms.h"
 #include "FileItem.h"
-#include "ViewState.h"
+#include "view/ViewState.h"
 #include "settings/GUISettings.h"
 #include "settings/Settings.h"
 #include "filesystem/Directory.h"
