+++ xbmc/pvr/dialogs/GUIDialogPVRChannelsOSD.h
@@ -20,7 +20,7 @@
  */
 
 #include "guilib/GUIDialog.h"
-#include "GUIViewControl.h"
+#include "view/GUIViewControl.h"
 #include "utils/Observer.h"
 #include "pvr/channels/PVRChannelGroupsContainer.h"
 #include <map>
