+++ xbmc/video/dialogs/GUIDialogVideoBookmarks.cpp
@@ -28,7 +28,7 @@
 #endif
 #include "pictures/Picture.h"
 #include "dialogs/GUIDialogContextMenu.h"
-#include "ViewState.h"
+#include "view/ViewState.h"
 #include "settings/Settings.h"
 #include "settings/AdvancedSettings.h"
 #include "FileItem.h"
