+++ xbmc/video/dialogs/GUIDialogVideoBookmarks.h
@@ -21,7 +21,7 @@
  */
 
 #include "guilib/GUIDialog.h"
-#include "GUIViewControl.h"
+#include "view/GUIViewControl.h"
 #include "video/VideoDatabase.h"
 
 class CFileItemList;
