+++ xbmc/view/GUIViewControl.cpp
