+++ xbmc/view/GUIViewState.h
