+++ xbmc/view/Makefile
@@ -1,0 +1,8 @@
+SRCS=GUIViewControl.cpp \
+     GUIViewState.cpp \
+     ViewDatabase.cpp \
+
+LIB=view.a
+
+include ../../Makefile.include
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
