+++ xbmc/view/ViewDatabase.cpp
@@ -21,7 +21,7 @@
 #include "ViewDatabase.h"
 #include "utils/URIUtils.h"
 #include "settings/Settings.h"
-#include "ViewState.h"
+#include "view/ViewState.h"
 #include "utils/log.h"
 #ifdef _LINUX
 #include "linux/ConvUtils.h" // GetLastError()
