+++ xbmc/view/ViewDatabase.h
