+++ xbmc/view/ViewState.h
